requirements.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.13.3
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
Flask-Cors==5.0.0
frozendict==2.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.0
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==1.24.4
openai==1.68.2
pandas==2.0.3
peewee==3.17.9
platformdirs==4.3.6
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing-extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
werkzeug==3.0.6
yfinance==0.2.55
zipp==3.20.2
